Add files via upload
This was one of my first projects I ever created. I followed a youtube video to create a text-based adventure game with some simple decision trees. Each decision had about two or three decisions after them based on the character that was chosen. This was a lot of fun! I really enjoyed learning javascript and now being able to look back at it.